% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_monobound.R
\name{alldup}
\alias{alldup}
\title{Auxiliary function: determine duplicates}
\usage{
alldup(data)
}
\arguments{
\item{data}{the data set on which to check for duplicate rows.}
}
\value{
boolean vector. Each entry indicates whether or not the
    corresponding row has a duplicate.
}
\description{
Auxiliary function that takes in a data set, and then determines
all the rows that are duplicates. It tags both the original row as
well as the duplicates.
}
