% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{altDefSplinesBasis}
\alias{altDefSplinesBasis}
\title{(Alternative) Defining single splines basis functions, with
interactions}
\usage{
altDefSplinesBasis(splineslist, j, l, v = 1)
}
\arguments{
\item{splineslist}{a list of splines commands and names of
variables that interact with the splines. This is generated
using the command \code{\link{removeSplines}}.}

\item{j}{the index for the spline for which to generate the basis
functions.}

\item{l}{the index for the basis.}

\item{v}{a constant that multiplies the spline basis.}
}
\value{
a vectorized function corresponding to a single splines
    basis function that can be numerically integrated.
}
\description{
This function returns a numerically integrable function
corresponding to a single splines basis function. It was not
implemented because it was slower than using the function from the
\code{splines2} package.
}
