% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callcheck.R
\name{argstring}
\alias{argstring}
\title{Auxiliary function: extract arguments from function in string form}
\usage{
argstring(string)
}
\arguments{
\item{string}{the function in string form.}
}
\value{
string of arguments.
}
\description{
Auxiliary function to extract arguments from a function that is in
string form.
}
