% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{bound.mst}
\alias{bound.mst}
\title{Obtaining TE bounds}
\usage{
bound.mst(g0, g1, sset, lpobj, obseq.tol, lpsolver, noisy = FALSE)
}
\arguments{
\item{g0}{set of expectations for each terms of the MTR for the
control group.}

\item{g1}{set of expectations for each terms of the MTR for the
control group.}

\item{sset}{a list containing the point estimates and gamma
components associated with each element in the S-set.}

\item{lpobj}{A list of matrices and vectors defining an LP problem.}

\item{obseq.tol}{tolerance level for how much more the solution is
permitted to violate observational equivalence of the IV-like
estimands.}

\item{lpsolver}{string, name of the package used to solve the LP
problem.}

\item{noisy}{boolean, set to \code{TRUE} if optimization results
should be displayed.}
}
\value{
a list containing the bounds on the treatment effect; the
    coefficients on each term in the MTR associated with the upper
    and lower bounds, for both counterfactuals; the optimization
    status to the maximization and minimization problems; the LP
    problem that the optimizer solved.
}
\description{
This function estimates the bounds on the target treatment effect.
}
