% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{bX}
\alias{bX}
\title{Spline basis function of order 1}
\usage{
bX(x, knots, i)
}
\arguments{
\item{x}{vector, the values at which to evaluate the basis
function.}

\item{knots}{vector, the internal knots.}

\item{i}{integer, the basis component to be evaluated.}
}
\value{
scalar.
}
\description{
This function is the splines basis function of order 1. This
function was coded in accordance to Carl de Boor's set of notes on
splines, "B(asic)-Spline Basics".
}
