% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{checkU}
\alias{checkU}
\title{Check polynomial form of the u-term}
\usage{
checkU(formula, uname)
}
\arguments{
\item{formula}{a formula.}

\item{uname}{name of the unobserved variable.}
}
\value{
If the unobservable term is entered correctly into the
    formula, then \code{NULL} is returned. Otherwise, the vector of
    incorrect terms is returned.
}
\description{
This function ensures that the unobservable term enters into the
MTR in the correct manner. That is, it enters as a polynomial.
}
