% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_callcheck.R
\name{class_formula}
\alias{class_formula}
\title{Auxiliary function: test if object is a formula}
\usage{
class_formula(obj)
}
\arguments{
\item{obj}{the object to be checked.}
}
\value{
boolean expression.
}
\description{
Auxiliary function to test if an object is a formula. Warnings are
suppressed.
}
