% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_callcheck.R
\name{class_list}
\alias{class_list}
\title{Auxiliary function: test if object is a list}
\usage{
class_list(obj)
}
\arguments{
\item{obj}{the object to be checked.}
}
\value{
boolean expression.
}
\description{
Auxiliary function to test if an object is a list. Warnings are
suppressed.
}
