% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monobound.R
\name{combinemonobound}
\alias{combinemonobound}
\title{Combining the boundedness and monotonicity constraint objects}
\usage{
combinemonobound(bdA, monoA)
}
\arguments{
\item{bdA}{list containing the constraint matrix, vector of
inequalities, and RHS vector associated with the boundedness
constraints.}

\item{monoA}{list containing the constraint matrix, vector on
inequalities, and RHS vector associated with the monotonicity
constraints.}
}
\value{
a list containing a unified constraint matrix, unified
    vector of inequalities, and unified RHS vector for the
    boundedness and monotonicity constraints of an LP problem.
}
\description{
This function simply combines the objects associated with the
boundedness constraints and the monotonicity constraints.
}
