% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{constructConstant}
\alias{constructConstant}
\title{Construct constant function}
\usage{
constructConstant(x)
}
\arguments{
\item{x}{scalar, the constant the function evaluates to.}
}
\value{
a function.
}
\description{
This function constructs another function that returns a
constant. It is used for constructing weight/knot functions.
}
