% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{defSplines}
\alias{defSplines}
\title{Defining list of splines basis functions, with interactions}
\usage{
defSplines(splineslist, j, v)
}
\arguments{
\item{splineslist}{a list of splines commands and names of
variables that interact with the splines. This is generated
using the command \code{\link{removeSplines}}.}

\item{j}{the index for the spline for which to generate the basis
functions.}

\item{v}{a constant that multiplies the spline basis.}
}
\value{
a list of vectorized functions corresponding to a single
    splines basis functions that can be numerically integrated.
}
\description{
This functions generates functions for each splines basis, and
returns it as a list.
}
