% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtb}
\alias{dtb}
\title{Basic data set distribution}
\format{A data frame with 12 rows and 7 columns.
\describe{
  \item{x}{covariate}
  \item{z}{instrument}
  \item{p}{probability of treatment uptake}
  \item{ey1}{counterfactual outcome when a recipient of treatment}
  \item{ey0}{counterfactual outcome when not a recipient of treatment}
  \item{f}{density}
  \item{multiplier}{number of observations in the data set}
}}
\source{
Simulated.
}
\usage{
dtb
}
\description{
The distribution from which \code{dtbf} was generated. The
distribution is defined by the (non-exported) command
\code{gendist_basic()}.  It is worth noting that the unobservable
terms are alreayd integrated out.
}
\keyword{datasets}
