% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtbf}
\alias{dtbf}
\title{Basic data set}
\format{An object of class \code{data.frame} with 1000 rows and 10 columns.}
\source{
Simulated.
}
\usage{
dtbf
}
\description{
A simulated population-level data involving only one covariate. The
data is constructed by the (non-exported) command
\code{gendist_basic()}.  It is worth noting that the unobservable
terms are alreayd integrated out.
}
\keyword{datasets}
