% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtc}
\alias{dtc}
\title{Covariates data set distribution}
\format{A data frame with 36 rows and 10 columns.
\describe{
  \item{x1}{covariate 1}
  \item{x2}{covariate 2}
  \item{z1}{instrument 1}
  \item{z2}{instrument 2}
  \item{latent}{latent variable determining treatment participation}
  \item{p}{probability of treatment uptake}
  \item{ey0}{counterfactual outcome when not a recipient of treatment}
  \item{ey1}{counterfactual outcome when a recipient of treatment}
  \item{f}{density}
  \item{multiplier}{number of observations required in data set}
}}
\source{
Simulated.
}
\usage{
dtc
}
\description{
The distribution from which \code{dtcf} was generated.
}
\keyword{datasets}
