% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtm}
\alias{dtm}
\title{Mosquito data set}
\format{A data frame with 400 rows and 7 columns.
\describe{
  \item{i}{index for observation}
  \item{z}{categorical variable for level of subsidy}
  \item{pz}{probability of purchasing mosquito net}
  \item{d}{indicator for whether or not mosquito net was purchased}

  \item{ey0}{counterfactual probability of contracting malaria
  conditional on not purchasing a mosquito net.}

  \item{ey1}{counterfactual probability of contracting malaria
  conditional on purchasing a mosquito net.}

  \item{ey}{the observed probability of contracting malaria.}
}}
\source{
Simulated, based on Mogstad, Torgovitsky (2017).
}
\usage{
dtm
}
\description{
A simulated population-level data set characterizing the effect of
purchasing mosquito nets, and the likelihood of contracting
malaria. A randomly assigned subsidy to purchase mosquito nets was
provided as part of an experiment.
}
\keyword{datasets}
