% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dts}
\alias{dts}
\title{Splines data set distribution}
\format{A data frame with 6 rows and 9 columns.
\describe{
  \item{group}{indicates combination of (X, Z)}
  \item{x}{covariate}
  \item{z}{instrument}
  \item{f}{density}
  \item{p}{probability of treatment uptake}
  \item{ey1}{counterfactual outcome when a recipient of treatment}
  \item{ey0}{counterfactual outcome when not a recipient of treatment}
  \item{multiplier}{number of observations in the data set}
  \item{controls}{number of controls in the data set}
}}
\source{
Simulated.
}
\usage{
dts
}
\description{
The distribution from which \code{dtsf} was generated.
}
\keyword{datasets}
