% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweights.R
\name{extractcols}
\alias{extractcols}
\title{Auxiliary function: extracting columns by component names}
\usage{
extractcols(M, components)
}
\arguments{
\item{M}{The matrix to extract from.}

\item{components}{The vector of variable names.}
}
\description{
Auxiliary function to extract columns from a matrix based on column
mames.
}
