% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{fmtResult}
\alias{fmtResult}
\title{Format result for display}
\usage{
fmtResult(x)
}
\arguments{
\item{x}{The scalar to be formated}
}
\value{
A scalar.
}
\description{
This function simply takes a number and formats it for being
displayed. Numbers less than 1 in absolute value are rounded to 6
significant figure. Numbers larger than
}
