% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{funAdd}
\alias{funAdd}
\title{Auxiliary function: add functions together}
\usage{
funAdd(FUN1, FUN2)
}
\arguments{
\item{FUN1}{the first function to be summed.}

\item{FUN2}{the second function to be summed.}
}
\value{
a function that is the sum of FUN1 and FUN2.
}
\description{
This function simply takes in two different functions with the same
arguments, and takes the sum of them.
}
