% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{funEval}
\alias{funEval}
\title{Evaluate a particular function}
\usage{
funEval(fun, values = NULL, argnames = NULL)
}
\arguments{
\item{fun}{the function to be integrated.}

\item{values}{the values of the arguments to the function.}

\item{argnames}{the argument names corresponding to \code{values}.}
}
\value{
the output of the function evaluated.
}
\description{
This function evaluates a single function in a list of functions.
}
