% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{funMultiply}
\alias{funMultiply}
\title{Auxiliary function: multiply functions together}
\usage{
funMultiply(FUN1, FUN2)
}
\arguments{
\item{FUN1}{the first function to be multiplied.}

\item{FUN2}{the second function to be multiplied.}
}
\value{
a function that is the product of FUN1 and FUN2.
}
\description{
This function simply takes in two different functions with the same
arguments, and takes the product of them.
}
