% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{gendist3}
\alias{gendist3}
\title{Generate test distribution 3}
\usage{
gendist3(subN = 5, p1 = 0.4, p2 = 0.6)
}
\arguments{
\item{subN}{integer, default set to 5. This is the number of
individuals possessing each value of the instrument. So the
total number of observations is subN * 2.}

\item{pi1}{the probability of treatment for those with the
instrument Z = 1.}

\item{pi2}{the probability of treatment for those with the
instrument Z = 2.}
}
\value{
a data.frame.
}
\description{
This function generates a data set for testing purposes. There is a
single instrument that takes on values of 1 and 2, and the
distribution of the values for the binary instrument is
uniform. The MTRs are m0 ~ 1 and m1 ~ 1. All unobservables u are
integrated out.
}
