% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{gendist4}
\alias{gendist4}
\title{Generate test distribution 4}
\usage{
gendist4(subN = 5, p1 = 0.4, p2 = 0.6, p3 = 0.8)
}
\arguments{
\item{subN}{integer, default set to 5. This is the number of
individuals possessing each value of the instrument. So the
total number of observations is subN * 2.}

\item{p1}{the probability of treatment for those with the
instrument Z = 1.}

\item{p2}{the probability of treatment for those with the
instrument Z = 2.}

\item{p3}{the probability of treatment for those with the
instrument Z = 3.}
}
\value{
a data.frame.
}
\description{
This function generates a data set for testing purposes. There is a
single instrument that takes on values of 1, 2, and 3, and the
distribution of the values for the binary instrument is
uniform. The MTRs are m0 ~ 1 and m1 ~ 1. All unobservables u are
integrated out.
}
