% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{gendist5e}
\alias{gendist5e}
\title{Generate test distribution 5 (has errors and a covariate)}
\usage{
gendist5e(N = 100, subN = 0.5, p1 = 0.4, p2 = 0.6, v0.sd = 1,
  v1.sd = 1.55)
}
\arguments{
\item{N}{integer, default set to 100. Total number of observations
in the data.}

\item{subN}{, default set to 0.5. This is the probability the agent
will have Z = 1.}

\item{p1}{the probability of treatment for those with the
instrument Z = 1.}

\item{p2}{the probability of treatment for those with the
instrument Z = 2.}

\item{v0.sd}{numeric, standard deviation of error term for
counterfactual \code{D = 0}}

\item{v1.sd}{numeric, standard deviation of error term for
counterfactual \code{D = 1}}
}
\value{
a data.frame.
}
\description{
This function generates a data set for testing purposes. There is a
single instrument that takes on values of 1 or 2, and the
distribution of the values for the binary instrument is
uniform. The MTRs are both of the form m ~ 1 + x + u.
}
