% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{gendist6e}
\alias{gendist6e}
\title{Generate test distribution 6 (has errors and a covariate)}
\usage{
gendist6e(N = 100, v0.sd = 1, v1.sd = 1.55)
}
\arguments{
\item{N}{integer, default set to 100. Total number of observations
in the data.}

\item{v0.sd}{numeric, standard deviation of error term for
counterfactual \code{D = 0}}

\item{v1.sd}{numeric, standard deviation of error term for
counterfactual \code{D = 1}}
}
\value{
a data.frame.
}
\description{
This function generates a data set for testing purposes. There is a
single instrument that is uniformly distributed over [0, 1]. The
MTRs are both of the form m ~ 1 + x + x:u.
}
