% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{gendist_basic}
\alias{gendist_basic}
\title{Generate basic data set for testing}
\usage{
gendist_basic()
}
\value{
a list of two data.frame objects. One is the distribution
    of the simulated data, the other is the full simulated data
    set.
}
\description{
This code generates population level data to test the estimation
function. This is a simpler dataset, one in which we can more
easily estimate a correctly specified model.  The data presented
below will have already integrated over the # unobservable terms
U, where U | X, Z ~ Unif[0, 1].
}
