% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{gendist_mosquito}
\alias{gendist_mosquito}
\title{Generate mosquito data set}
\usage{
gendist_mosquito()
}
\value{
data.frame.
}
\description{
This code generates the population level data in Mogstad, Santos,
Torgovitsky (2018), i.e. the mosquito data set used as the running
example.
}
