% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{gendistCovariates}
\alias{gendistCovariates}
\title{Generate test data set with covariates}
\usage{
gendistCovariates()
}
\value{
a list of two data.frame objects. One is the distribution
    of the simulated data, the other is the full simulated data
    set.
}
\description{
This code generates population level data to test the estimation
function. This data includes covariates. The data generated will
have already integrated over the unobservable terms U, where U | X,
Z ~ Unif[0, 1].
}
