% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{gendistMosquito}
\alias{gendistMosquito}
\title{Generate mosquito data set}
\usage{
gendistMosquito()
}
\value{
data.frame.
}
\description{
This code generates the population level data in Mogstad, Santos,
Torgovitsky (2018), i.e. the mosquito data set used as the running
example.
}
