% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweights.R
\name{genej}
\alias{genej}
\title{Auxiliary function: generating basis vectors}
\usage{
genej(pos, length)
}
\arguments{
\item{pos}{The position of the non-zero entry/dimension the basis
vector corresponds to}

\item{length}{Number of dimensions in total/length of vector.}
}
\value{
Vector containing 1 in a single position, and 0 elsewhere.
}
\description{
Auxiliary function to generate standard basis vectors.
}
