% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monobound.R
\name{genfullmbA}
\alias{genfullmbA}
\title{Combining the boundedness and monotonicity constraint objects}
\usage{
genfullmbA(bdA, monoA)
}
\arguments{
\item{bdA}{list containing the constraint matrix, vector of
inequalities, and RHS vector associated with the boundedness
constraints.}

\item{monoA}{list containing the constraint matrix, vector on
inequalities, and RHS vector associated with the monotonicity
constraints.}
}
\value{
a list containing a unified constraint matrix, unified
    vector of inequalities, and unified RHS vector for the
    boundedness and monotonicity constraints.
}
\description{
This function combines the objects associated with the boundedness
constraints and the monotonicity constraints.
}
