% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{gensset.mst}
\alias{gensset.mst}
\title{Generating LP moments for IV-like estimands}
\usage{
gensset.mst(data, sset, sest, splinesobj, pmodobj, pm0, pm1, ncomponents,
  scount, subset_index)
}
\arguments{
\item{sset}{A list, which is modified and returned as the output.}

\item{sest}{A list containing the point estimates and S-weights
corresponding to a particular IV-like estimand.}

\item{pmodobj}{A vector of propensity scores.}

\item{pm0}{A list of the monomials in the MTR for d = 0.}

\item{pm1}{A list of the monomials in the MTR for d = 1.}

\item{ncomponents}{The number of components from the IV regression
we want to include in the S-set.}

\item{scount}{A counter for the number of elements in the S-set.}

\item{subset_index}{An index for the subset of the data the IV
regression is restricted to.}
}
\value{
A list containing the point estimate for the IV regression,
    and the expectation of each monomoial term in the MTR.
}
\description{
This function takes in the IV estimate and its IV-like
specification, and generates a list containing the corresponding
point estimate, and the corresponding moments (gammas) that will
enter into the contraint matrix of the LP problem.
}
