% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{gmmEstimate}
\alias{gmmEstimate}
\title{GMM estimate of TE under point identification}
\usage{
gmmEstimate(sset, gstar0, gstar1, itermax = 2, tol = 1e-08,
  noisy = TRUE)
}
\arguments{
\item{sset}{a list of lists constructed from the function
\link{gensset}. Each inner list should include a
coefficient corresponding to a term in an IV specification, a
matrix of the estimates of the gamma moments conditional on (X,
Z) for d = 0, and a matrix of the estimates of the gamma
moments conditional on (X, Z) for d = 1. The column means of
the last two matrices is what is used to generate the gamma
moments.}

\item{gstar0}{vector, the target gamma moments for d = 0.}

\item{gstar1}{vector, the target gamma moments for d = 1.}

\item{itermax}{integer, maximum number of iterations allowed in the
iterative GMM process. By default this is set to 2 (two-step
GMM).}

\item{tol}{tolerance level for iterative GMM to terminate.}

\item{noisy}{boolean, default set to \code{TRUE}. If \code{TRUE},
then messages are provided throughout the estimation
procedure. Set to \code{FALSE} to suppress all messages,
e.g. when performing the bootstrap.}
}
\value{
a list containing the point estimate of the treatment
    effects, the standard errors, the 90% and 95% confidence
    intervals, the convergence code (see
    \code{\link[stats]{optim}}), the coefficients on the MTR, and
    the variance/covariance matrix of the MTR coefficient
    estimates.
}
\description{
If the user sets the argument \code{point = TRUE} in the function
\code{ivmte}, then it is assumed that the treatment effect
parameter is point identified. The observational equivalence
condition is then set up as a GMM problem. Solving this GMM problem
recovers the coefficients on the MTR functions m0 and m1. Combining
these coefficients with the target gamma moments allows us to
estimate the target treatment effect.
}
