% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_callcheck.R
\name{isfunctionstring}
\alias{isfunctionstring}
\title{Auxiliary function: check if string is command}
\usage{
isfunctionstring(string)
}
\arguments{
\item{string}{the string object to be checked.}
}
\value{
boolean expression.
}
\description{
Auxiliary function to check if a string is in fact a command, but
in string form.
}
