% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{l}
\alias{l}
\title{Listing subsets and components}
\usage{
l(...)
}
\arguments{
\item{...}{subset conditions or variable names}
}
\value{
list.
}
\description{
This function allows the user to declare a list of variable names
in non-character form and subsetting conditions. This is used to
ensure clean entry of arguments into the \code{components} and
\code{subset} arguments of the function.
}
\examples{
components <- l(d, x1, intercept)
subsets <- l(, z \%in\% c(2, 4))

}
