% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{listFunEval}
\alias{listFunEval}
\title{Evaluate a particular function in a list of functions}
\usage{
listFunEval(fun, values = NULL, argnames = NULL)
}
\arguments{
\item{values}{the values of the arguments to the function.}

\item{argnames}{the argument names corresponding to \code{values}.}

\item{funList}{the list of functions.}

\item{pos}{the index/position of the function to be evaluated in
\code{funList}.}
}
\value{
the output of the function evaluated.
}
\description{
This function evaluates a single function in a list of functions.
}
