% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{listIntegrate}
\alias{listIntegrate}
\title{Auxiliary funtion: integrate a list of functions}
\usage{
listIntegrate(list, subdivisions, rel.tol, abs.tol)
}
\arguments{
\item{list}{the list of functions to be multiplied.}
}
\value{
a list of scalars.
}
\description{
This function integrates each element of a list of functions using
numerical quadrature.
}
