% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{listMean}
\alias{listMean}
\title{Auxiliary function: mean of a list of numerically integrated
functions}
\usage{
listMean(integratedList, component)
}
\arguments{
\item{integratedList}{the list of scalars to be averaged over. The
format of each element is assumed to be that of the output from
R's built-in command \code{integrate}.}

\item{component}{the elements of the inner list for which one
wishes to obtain the average of.}
}
\value{
a scalar.
}
\description{
This function takes in a list of lists of scalars. It then
calculates the averages across the corresponding elements across
the lists.
}
