% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{listMultiply}
\alias{listMultiply}
\title{Auxiliary function: multiply a list of functions by a single function}
\usage{
listMultiply(list, multiplier)
}
\arguments{
\item{list}{the list of functions to be multiplied.}

\item{multiplier}{the function that is to multiply each element in
the list.}
}
\value{
a list of functions.
}
\description{
This function multiplies each element of a list of functions by a
single function.
}
