% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{lists.mst}
\alias{lists.mst}
\title{Listing subsets and components}
\usage{
lists.mst(...)
}
\arguments{
\item{...}{subset conditions or variable names}
}
\value{
list.
}
\description{
This function allows the user to declare a list of subsetting
conditions and variable names in non-character form. This is used
to ensure clean entry of arguments into the function.
}
\examples{
components <- lists.mst(d, x1, intercept)
subsets <- lists.mst(, z \%in\% c(2, 4))

}
