% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lpsetup.mst}
\alias{lpsetup.mst}
\title{Constructing LP problem}
\usage{
lpsetup.mst(sset, mbA = NULL, mbs = NULL, mbrhs = NULL, lpsolver,
  shape = TRUE)
}
\arguments{
\item{sset}{List of IV-like estimates and the corresponding gamma
terms.}

\item{mbA}{Matrix used to define the constraints in the LP problem.}

\item{mbs}{Vector of model sense/inequalities signs used to define
the constraints in the LP problem.}

\item{mbrhs}{Vector of constants used to define the constraints in
the LP problem.}

\item{lpsolver}{string, name of the package used to solve the LP
problem.}

\item{shape}{boolean, default set to TRUE. Switch to determine
whether or not to include shape restrictions in the LP problem.}

\item{bdASeq}{integer vector, row indexes in the \code{mbA} object
pertaining to boundedness restrictions.}

\item{monoASeq}{integer vector, row indexes in the \code{mbA} object
pertaining to monotonicity restrictions.}
}
\value{
A list of matrices and vectors necessary to define an LP
    problem for Gurobi.
}
\description{
This function takes in the IV estimates from the set of IV
regressions declared by the user, as well as their corresponding
moments of the terms in the MTR. These are then used to construct
the components that make up the LP problem. Additional constraint
matrix is added using \code{mbA} (\code{mb} stands for
"monotonicity/boundedness"); extra model sense is added using
\code{mbs}; extra RHS values added using \code{mbrhs}). Depending
on the linear programming solver used, this function will return
different output specific to the solver.
}
