% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lpSetupCriterion}
\alias{lpSetupCriterion}
\title{Configure LP environment for minimizing the criterion}
\usage{
lpSetupCriterion(env, sset)
}
\arguments{
\item{env}{The LP environment}

\item{sset}{List of IV-like estimates and the corresponding gamma
terms.}
}
\value{
Nothing, as this modifies an environment variable to save
    memory.
}
\description{
This function sets up the objective function for minimizing the
criterion.
}
