% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lpSetupInfeasible}
\alias{lpSetupInfeasible}
\title{Configure LP environment for diagnostics}
\usage{
lpSetupInfeasible(env, sset)
}
\arguments{
\item{env}{The LP environment}

\item{sset}{List of IV-like estimates and the corresponding gamma
terms.}
}
\value{
Nothing, as this modifies an environment variable to save
    memory.
}
\description{
This function separates the shape contraints from the LP
environment. That way, the model can be solved without any shape
constraints. This is done in order to check which shape constraints
are causing the model to be infeasible.
}
