% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lpSetupSolver}
\alias{lpSetupSolver}
\title{Configure LP environment to be compatible with solvers}
\usage{
lpSetupSolver(env, lpsolver)
}
\arguments{
\item{env}{The LP environment}

\item{lpsolver}{Character, the LP solver.}
}
\value{
Nothing, as this modifies an environment variable to save
    memory.
}
\description{
This alters the LP environment so the model will be compatible with
specific solvers.
}
