% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{m.int}
\alias{m.int}
\title{Function to generate integral of m0 and m1}
\usage{
m.int(ub, lb, coef)
}
\arguments{
\item{ub}{scalar, upper bound of the integral.}

\item{lb}{scalar, lower bound of the integral.}

\item{coef}{vector, polynomial coefficients.}
}
\value{
scalar.
}
\description{
Function carries out integral for a polynomial of degree 3.
}
