% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{magnitude}
\alias{magnitude}
\title{Check magnitude of real number}
\usage{
magnitude(x)
}
\arguments{
\item{x}{The number to be checked.}
}
\value{
An integer indicating the order of magnitude.
}
\description{
This function returns the order of magnitude of a a number.
}
