% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{matrixTriplets}
\alias{matrixTriplets}
\title{Convert matrix into triplet form}
\usage{
matrixTriplets(mat, lower = TRUE)
}
\arguments{
\item{mat}{A matrix.}

\item{lower}{Boolean, set to \code{TRUE} if matrix is symmetric,
and only its lower triangle should be returned.}
}
\value{
A list containing vectors of row and column indexes, and
    matrix values.
}
\description{
This function converts matrices into triplet form for Mosek.  This
is required in order to declare quadratic programming problems and
second-order cone programming problems.
}
