% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modcall.R
\name{modcall}
\alias{modcall}
\title{Auxiliary function: modifying calls}
\usage{
modcall(call, newcall, newargs, keepargs, dropargs, testing = FALSE)
}
\arguments{
\item{call}{Call object to be modified.}

\item{newcall}{New function to be called.}

\item{newargs}{List, new arguments and their values.}

\item{keepargs}{List, arguments in original call to keep, with the
rest being dropped.}

\item{dropargs}{List, arguments in original call to drop, with the
rest beingkept.}
}
\value{
New call object.
}
\description{
This function can be used to modify calls in several ways.
}
\examples{
modcall(call,
        newcall = propensity.mst,
        keepargs = c("link", "late.Z", "late.X"),
        dropargs = "propensity.mst",
        newargs = list(data = quote(cdata),
        formula = propensity))

}
