% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{monoIntegral}
\alias{monoIntegral}
\title{Integrating and evaluating monomials}
\usage{
monoIntegral(u, exp)
}
\arguments{
\item{u}{scalar, the point at which to evaluate the integral. If a
vector is passed, then the integral is evaluated at all the
elements of the vector.}

\item{exp}{The exponent of the monomial.}
}
\value{
scalar or vector, depending on what \code{u} is.
}
\description{
Analytically integrates monomials and evalates them at a given
point. It is assumed that there is no constant multiplying the
monomial.
}
