% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{mst}
\alias{mst}
\title{Estimation procedure from Mogstad, Torgovitsky (2017)}
\usage{
mst(formula, data, subset, components = NULL, propensity, link, treat, m0, m1,
  uname = u, target, late.Z, late.from, late.to, late.X, eval.X, wald.lb,
  wald.ub, threshold = 1e-08, u.n = 10, X.n = 10, add.audit = 2,
  max.audits = 5, m1.ub, m0.ub, m1.lb, m0.lb, mte.ub, mte.lb, m0.dec, m0.inc,
  m1.dec, m1.inc, mte.dec, mte.inc)
}
\arguments{
\item{formula}{formula or vector of formulas used to specify the
regressions for the IV-like estimands.}

\item{data}{\code{data.frame} used to estimate the treatment
effects.}

\item{subset}{single subset condition or list of subset conditions
correpsonding to each IV-like estimand. See
\code{\link[mst]{list.mst}} on how to input the argument.}

\item{components}{a list of vectors of the terms/components from
the regressions specifications we want to include in the set of
IV-like estimands.  See \code{\link[mst]{list.mst}} on how to
input the argument.}

\item{propensity}{formula or variable name corresponding to
propensity to take up treatment. If a formula is declared, then
the function estimates propensity score according to the
formula and link specified. If a variable name is declared,
then the corresponding column in the data is taken as the
vector of propensity scores.}

\item{link}{name of link function to estimate propensity score. Can
be chosen from \code{linear}, \code{probit}, or \code{logit}.}

\item{treat}{variable name for treatment indicator}

\item{m0}{one-sided formula for marginal treatment response
function for control group.}

\item{m1}{one-sided formula for marginal treatment response
function for treated group.}

\item{uname}{variable name for unobservale used in declaring MTRs.}

\item{target}{target parameter to be estimated. Currently function
allows for ATE ("ate"), ATT ("att"), ATU ("atu"), LATE
("late"), and generalized LATE ("genlate").}

\item{late.Z}{vector of variable names used to define the LATE.}

\item{late.from}{baseline set of values of Z used to define the
LATE.}

\item{late.to}{comparison set of values of Z used to define the
LATE.}

\item{late.X}{vector of variable names of covariates which we
condition on when defining the LATE.}

\item{eval.X}{numeric vector of the values at which we condition
variables in \code{late.X} on when estimating the LATE.}

\item{wald.lb}{lower bound value of unobservable u for estimating
generalized LATE.}

\item{wald.ub}{upper bound value of unobservable u for estimating
generalized LATE.}

\item{threshold}{threshold for violation of observational
equivalence.}

\item{u.n}{number of evenly spread points in the interval [0, 1] of
the unobservable u used to form the grid in the audit
procedure.}

\item{X.n}{number of evenly spread points of the covariates to use
to form the grid in the audit procedure.}

\item{add.audit}{number of points to add to the grid in each
iteration of the audit procedure.}

\item{max.audits}{maximum number of iterations in the audit
procedure.}

\item{m1.ub}{numeric value for upper bound on MTR for treated
group.}

\item{m0.ub}{numeric value for upper bound on MTR for control
group.}

\item{m1.lb}{numeric value for lower bound on MTR for treated
group.}

\item{m0.lb}{numeric value for lower bound on MTR for control
group.}

\item{mte.ub}{numeric value for upper bound on treatment effect
paramter of interest.}

\item{mte.lb}{numeric value for lower bound on treatment effect
paramter of interest.}

\item{m0.dec}{logical, equal to TRUE if we want MTR for control
group to be weakly monotone decreasing.}

\item{m0.inc}{logical, equal to TRUE if we want MTR for control
group to be weakly monotone increasing.}

\item{m1.dec}{logical, equal to TRUE if we want MTR for treated
group to be weakly monotone decreasing.}

\item{m1.inc}{logical, equal to TRUE if we want MTR for treated
group to be weakly monotone increasing.}

\item{mte.dec}{logical, equal to TRUE if we want the MTE to be
weakly monotone decreasing.}

\item{mte.inc}{logical, equal to TRUE if we want the MTE to be
weakly monotone decreasing.}
}
\value{
Returns a list of results from throughout the estimation
    procedure. This includes all IV-like estimands; the propensity
    score model; bounds on the treatment effect; the estimated
    expectations of each term in the MTRs; the components and
    results of the LP problem.
}
\description{
This function estimates bounds on treatment effect parameters,
following the procedure described in Mogstad, Torgotvitsky
(2017). Of the target parameters, the user can choose from the ATE,
ATT, ATU, LATE, and generalized LATE. The user is required to
provide a polynomial expression for the marginal treatment
responses (MTR), as well as a set of regressions. By restricting
the set of coefficients on each term of the MTRs to be consistent
with the regression estimates, the function is able to restrict
itself to a set of MTRs. The bounds on the treatment effect
parameter correspond to finding coefficients on the MTRs that
maximize their average difference.
}
\details{
The estimation procedure relies on the propensity to take up
treatment. The propensity scores can either be estimated as part of
the estimation procedure, or the user can specify a variable in the
data set already containing the propensity scores.

Constraints on the shape of the MTRs and marginal treatment effects
(MTE) can be imposed by the user, also. Specifically, bounds and
monotonicity restrictions are permitted. These constraints are only
enforced over a subset of the data. However, an audit procedure
randomly selects points outside of this subset to determine whether
or not the constraints hold. The user can specify how stringent
this audit procedure is using the function arguments.
}
\examples{
formulas <- c(ey ~ d | z,
              ey ~ d | factor(z),
              ey ~ d,
              ey ~ d | factor(z))
jvec <- lists.mst(d, d, d, d)
svec <- lists.mst(, , , z \%in\% c(2, 4))

mst(formula = tform, data = dtm,
    components = jvec, 
    propensity = pz,
    subset = svec,
    m0 = ~  u + u^2,
    m1 = ~  u + u^2,
    uname = u,
    target = "att",
    m0.dec = TRUE,
    m1.dec = TRUE)


mst(formula = y ~ d + x1 + x2 | x1 + x2 + z1 + z2, dt,
    components = d,
    propensity = d ~ z1 + z2 + x1 + x2,
    link = "logit",
    m0 = ~ x1 + x2,
    m1 = ~ x1  ,
    target = "late",
    Z = c(z1, z2),
    from = c(0,1),
    to= c(1,2))

}
