% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{obseqmin.mst}
\alias{obseqmin.mst}
\title{Minimizing violation of observational equivalence}
\usage{
obseqmin.mst(sset, lpobj)
}
\arguments{
\item{sset}{A list of IV-like estimates and the corresponding gamma
terms.}

\item{lpobj}{A list of matrices and vectors defining an LP problem.}
}
\value{
A list including the minimum violation of observational
    equivalence, the solution to the LP problem, and the status of
    the solution.
}
\description{
Given a set of IV-like estimates and the set of matrices/vectors
defining an LP problem, this function minimizes the violation of
observational equivalence under the L1 norm.
}
