% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweights.R
\name{olsj.mst}
\alias{olsj.mst}
\title{OLS weights}
\usage{
olsj.mst(X, components, treat)
}
\arguments{
\item{X}{Matrix of covariates, including the treatment indicator.}

\item{components}{Vector of variable names of which user wants the
S-weights for.}

\item{treat}{Variable name for the treatment indicator.}
}
\value{
A list of two vectors: one is the weight for D = 0, the
    other is the weight for D = 1.
}
\description{
Function generating the S-weights for OLS estimand, with controls.
}
