% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{optionsCplexAPISingle}
\alias{optionsCplexAPISingle}
\title{Function to parse a single set of options for CPLEX}
\usage{
optionsCplexAPISingle(name, vector)
}
\arguments{
\item{name}{string, name of the \code{cplexapi} function to call to
implement the option.}

\item{vector}{a named vector, contains the argument names and
values of the options. The \code{env} argument in the
\code{cplexapi} documentation should always be omitted.}
}
\value{
string, the command to be evaluated to implement a single
    option.
}
\description{
This function constructs a string to be parsed when \code{lpsolver}
is set to \code{cplexapi}.
}
