% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{optionsCplexAPITol}
\alias{optionsCplexAPITol}
\title{Function to extract feasibility tolerance from CPLEX options}
\usage{
optionsCplexAPITol(options)
}
\arguments{
\item{options}{list, the set of options submitted by the user.}
}
\value{
scalar, the level to set the audit tolerance at.
}
\description{
This function parses through the user-submitted CPLEX options to
determine what the feasibility tolerance is. This tolerance can
then be used for the audit.  If the user does not set the CPLEX
feasibility tolerance, then a default value of \code{1e-06} is
returned.
}
