% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{optionsLpSolveAPI}
\alias{optionsLpSolveAPI}
\title{Function to parse options for lp_solve}
\usage{
optionsLpSolveAPI(options)
}
\arguments{
\item{options}{list. The name of each item must be the name of the
option, and is case sensitive. The value assigned to each item
is the value to set the option to. The \code{lprec} argument
should always be omitted.}
}
\value{
string, the command to be evaluated to implement the
    options.
}
\description{
This function constructs a list of options to be parsed when
\code{solver} is set to \code{lpsolveapi}. The options permitted
are those that can be set via \code{lpSolveAPI::lp.control}, and
should be passed as a named list (e.g. \code{list(epslevel =
"tight")}).
}
