% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{parenthBoolean}
\alias{parenthBoolean}
\title{Correct boolean expressions in terms lists}
\usage{
parenthBoolean(termsList)
}
\arguments{
\item{termsList}{character vector, the vector of terms.}
}
\value{
character vector.
}
\description{
This function takes a vector of terms and places parentheses around
boolean expressions.
}
