% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{permute}
\alias{permute}
\title{Auxiliary function: generate all permutations of a vector}
\usage{
permute(vector)
}
\arguments{
\item{vector}{The vector whose elements are to be permuted.}
}
\value{
a list of all the permutations of \code{vector}.
}
\description{
This function generates every permutation of the elements in a
vector.
}
