% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{permuteN}
\alias{permuteN}
\title{Auxiliary function: generate all permutation orderings}
\usage{
permuteN(n)
}
\arguments{
\item{n}{integer, the first n natural numbers one wishes to
permute.}
}
\value{
a list of all the permutations of the first n natural
    numbers.
}
\description{
This function generates every permutation of the first n natural
numbers.
}
