% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{piv.mst}
\alias{piv.mst}
\title{Obtaining IV-like estimands}
\usage{
piv.mst(Y, X, Z, lmcomponents, weights = NULL)
}
\arguments{
\item{Y}{the vector of outcomes.}

\item{X}{the matrix of covariates (includes endogenous and
exogenous covariates).}

\item{Z}{the matrix of instruments (includes exogenous covariates
in the second stage).}

\item{lmcomponents}{vector of variable names from the second stage
that we want to include in the S-set of IV-like estimands.}

\item{weights}{vector of weights.}

\item{formula}{the regression formula. This is required only so
that informative error messages can be returned to the user.}
}
\value{
vector of select coefficient estimates.
}
\description{
This function performs TSLS to obtain the estimates for the IV-like
estimands.
}
\examples{
piv.mst(y ~ x, Y, X, Z, c(d, x1, x2))
}
