% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{piv}
\alias{piv}
\title{Obtaining IV-like estimands}
\usage{
piv(Y, X, Z, lmcomponents = NULL, weights = NULL, order = NULL,
  excluded = TRUE)
}
\arguments{
\item{Y}{the vector of outcomes.}

\item{X}{the matrix of covariates (includes endogenous and
exogenous covariates).}

\item{Z}{the matrix of instruments (includes exogenous covariates
in the second stage).}

\item{lmcomponents}{vector of variable names from the second stage
that we want to include in the S-set of IV-like estimands. If
\code{NULL} is submitted, then all components will be included.}

\item{weights}{vector of weights.}

\item{order}{integer, the counter for which IV-like specification
and component the regression is for.}

\item{excluded}{boolean, to indicate whether or not the regression
involves excluded variables.}
}
\value{
vector of select coefficient estimates.
}
\description{
This function performs TSLS to obtain the estimates for the IV-like
estimands.
}
