% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{polylisteval}
\alias{polylisteval}
\title{Evaluating polynomials}
\usage{
polylisteval(polynomials, points)
}
\arguments{
\item{polynomials}{a list of polynomials.}

\item{points}{a list/vector of points at which we want to evaluate
each polynomial.}
}
\value{
A matrix of values, corresponding to the polynomials
    specified in \code{polynomials} evaluated at the points
    specified in \code{points}.
}
\description{
This function allows one to evaluate a list of polynomials at
various points (the points are allowed to differ across
polynomials). This function will instead be used to evaluate a list
of monomials.
}
