% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{polyparse.mst}
\alias{polyparse.mst}
\title{Parsing marginal treatment response formulas}
\usage{
polyparse.mst(formula, data, uname = u)
}
\arguments{
\item{formula}{the MTR.}

\item{data}{\code{data.frame} for which we obtain E[md | D, X, Z] for each
observation.}

\item{uname}{variable name for unobservable used in declaring the
MTR.}
}
\value{
A list (of lists) of monomials corresponding to the
    original MTR (for each observation); a list (of lists) of the
    integrated monomials; a vector for the degree of each of the
    original monomials in the MTR; and a vector for the names of
    each variable entering into the MTR (note \code{x^2 + x} has
    only one term, \code{x}).
}
\description{
This function takes in an MTR formula, and then parses the formula
such that it becomes a polynomial in the unobservable u. It then
breaks these polynomials into monomials, and then integrates each
of them with respect to u. Each integral corresponds to E[md | D,
X, Z].
}
