% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{polyProduct}
\alias{polyProduct}
\title{Function to multiply polynomials}
\usage{
polyProduct(poly1, poly2)
}
\arguments{
\item{poly1}{vector, characerizing a polynomial.}

\item{poly2}{vector, characerizing a polynomial.}
}
\value{
vector, characterizing the product of the two polynomials
    characterized \code{poly1} and \code{poly2}.
}
\description{
This function takes in two vectors characterizing polynomials. It
then returns a vector characterizing the product of the two
polynomials.
}
