% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{print.ivmte}
\alias{print.ivmte}
\title{Print results}
\usage{
\method{print}{ivmte}(x, ...)
}
\arguments{
\item{x}{an object returned from '\code{ivmte}'.}

\item{...}{additional arguments.}
}
\value{
basic set of results.
}
\description{
This function uses the print method on the ivmte return list.
}
