% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{qpSetup}
\alias{qpSetup}
\title{Constructing QCQP problem}
\usage{
qpSetup(env, sset, rescale = TRUE)
}
\arguments{
\item{env}{environment containing the matrices defining the LP
problem.}

\item{sset}{A list containing the covariats and outcome variable
for the direct MTR regression.}

\item{rescale}{boolean, set to \code{TRUE} if the MTR components
should be rescaled to improve stability in the LP/QP/QCP
optimization.}
}
\description{
This function is only used when the direct MTR regression procedure
is used. This function simply constructs the quadratic constraint,
and adds it to the LP problem defined by the linear optimization problem
for the bounds and the linear shape constraints.
}
