% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{qpSetupCriterion}
\alias{qpSetupCriterion}
\title{Configure QCQP problem to find minimum criterion}
\usage{
qpSetupCriterion(env)
}
\arguments{
\item{env}{The LP environment}
}
\value{
Nothing, as this modifies an environment variable to save
    memory.
}
\description{
This function sets up the objective function for minimizing the
criterion. The QCQP model must be passed as an environment variable,
under the entry \code{$model}. See \code{\link{qpSetup}}.
}
