% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{rescaleX}
\alias{rescaleX}
\title{Function to implement rescaling procedure}
\usage{
rescaleX(sset, dVec, drY, drN, shape.constraint = FALSE)
}
\arguments{
\item{sset}{a list of lists constructed from the function
\link{genSSet}. In the case of a direct regression, 'sset'
contains only one inner list. This list contains the gamma
moment at the individual level.}
}
\description{
This function rescales the matrix of covariates used in the direct
regression to improve the conditioning number and the stability of
the estimation procedure.
}
