% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wweights.R
\name{restring}
\alias{restring}
\title{Auxiliary function that converts an expression of variable names
into a vector of strings.}
\usage{
restring(vector, substitute = TRUE)
}
\arguments{
\item{vector}{An expression of a list of variable names.}

\item{substitute}{Boolean option of whether or not we wish to use
the \code{substitute} command when implementing this
function. Note that this substitutes the argument of the
function. If \code{substitute = FALSE}, then the function will
instead treat the arguments as variables, and substitute in
their values.}
}
\value{
A vector of variable names (strings).
}
\description{
Auxiliary function that converts an expression of variable names
into a vector of strings.
}
\examples{
a <- 4
b <- 5
ivmte:::restring(c(a, b), substitute = TRUE)
ivmte:::restring(c(a, b), substitute = FALSE)
}
