% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{rhalton}
\alias{rhalton}
\title{Generate Halton sequence}
\usage{
rhalton(n, base = 2)
}
\arguments{
\item{n}{Number of draws.}

\item{base}{Base used for the Halton sequence, set to 2 by default.}
}
\value{
A sequence of randomly drawn numbers.
}
\description{
This function generates a one dimensional Halton sequence.
}
