% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{runCplexAPI}
\alias{runCplexAPI}
\title{Running cplexAPI LP solver}
\usage{
runCplexAPI(lpobj, lpdir)
}
\arguments{
\item{lpobj}{list of matrices and vectors defining the linear
programming problem.}

\item{lpdir}{input either CPX_MAX or CPX_MIN, which sets the LP
problem as a maximization or minimization problem.}
}
\value{
a list of the output from CPLEX. This includes the
    optimization status, the objective value, the solution vector,
    amongst other things.
}
\description{
This function solves the LP problem using the cplexAPI package. The
object generated by \code{\link{lpsetup.mst}} is not compatible
with the \code{cplexAPI} functions. This function adapts the object
to solve the LP problem.
}
