% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{runLpSolveAPI}
\alias{runLpSolveAPI}
\title{Running lpSolveAPI}
\usage{
runLpSolveAPI(lpobj, modelsense, lpsolver.options)
}
\arguments{
\item{lpobj}{list of matrices and vectors defining the linear
programming problem.}

\item{modelsense}{input either 'max' or 'min' which sets the LP
problem as a maximization or minimization problem.}

\item{lpsolver.options}{list, each item of the list should
correspond to an option specific to the LP solver
selected.}
}
\value{
a list of the output from \code{lpSolveAPI}. This includes
    the optimization status, the objective value, the solution
    vector.
}
\description{
This function solves the LP problem using the \code{lpSolveAPI}
package. The object generated by \code{\link{lpSetup}} is not
compatible with the \code{lpSolveAPI} functions. This function
adapts the object to solve the LP problem.
}
