% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{runMosek}
\alias{runMosek}
\title{Running Rmosek}
\usage{
runMosek(lpobj, modelsense, solver.options, debug = FALSE)
}
\arguments{
\item{lpobj}{list of matrices and vectors defining the linear
programming problem.}

\item{modelsense}{input either 'max' or 'min' which sets the LP
problem as a maximization or minimization problem.}

\item{solver.options}{list, each item of the list should
correspond to an option specific to the LP solver selected.}
}
\value{
a list of the output from \code{lpSolveAPI}. This includes
    the objective value, the solution vector, and the optimization
    status (status of \code{1} indicates successful optimization).
}
\description{
This function solves the LP problem using the \code{Rmosek}
package. The object generated by \code{\link{lpSetup}} is not
compatible with the \code{Rmosek} functions. This function
adapts the object to solve the LP problem. See
\code{\link{runGurobi}} and \code{\link{runCplexAPI}} for
additional error code labels.
}
