% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{s.ols1.d}
\alias{s.ols1.d}
\title{IV-like weighting function, OLS specification 1}
\usage{
s.ols1.d(d)
}
\arguments{
\item{d}{0 or 1, indicating treatment or control.}
}
\value{
scalar.
}
\description{
IV-like weighting function for OLS specification 1.
}
