% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{s.ols2.d}
\alias{s.ols2.d}
\title{IV-like weighting function, OLS specification 2}
\usage{
s.ols2.d(x, d, exx)
}
\arguments{
\item{x}{vector, the value of the covariates other than the
intercept and the treatment indicator.}

\item{d}{0 or 1, indicating treatment or control.}

\item{exx}{the matrix E[XX']}
}
\value{
scalar.
}
\description{
IV-like weighting function for OLS specification 2.
}
