% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{s.tsls}
\alias{s.tsls}
\title{IV-like weighting function, TSLS specification}
\usage{
s.tsls(z, j)
}
\arguments{
\item{z}{vector, the value of the instrument.}

\item{j}{scalar, position of the component one is interested in
constructing the IV-like weight for.}
}
\value{
scalar.
}
\description{
IV-like weighting function for TSLS specification.
}
