% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{s.wald}
\alias{s.wald}
\title{IV-like weighting function, Wald specification}
\usage{
s.wald(z)
}
\arguments{
\item{z}{vector, the value of the instrument.}
}
\value{
scalar.
}
\description{
IV-like weighting function for OLS specification 2.
}
