% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{selectViolations}
\alias{selectViolations}
\title{Select points from audit grid to add to the initial grid}
\usage{
selectViolations(diffVec, audit.add, lb0seq, lb1seq, ub0seq, ub1seq,
  mono0seq, mono1seq, monoteseq, mbmap)
}
\arguments{
\item{diffVec}{numeric vector, with a positive value indicating a
violation of a shape constriant.}

\item{audit.add}{integer, the number of points from the audit grid
to add to the intial for each constraint type. For instance, if
there are 5 different kinds of constraints imposed, and
\code{audit.add = 5}, then up to 30 points may be added to the
initial grid.}

\item{lb0seq}{integer vector, indicates which rows in the audit
constraint matrix corresponding to the lower bound for m0.}

\item{lb1seq}{integer vector, indicates which rows in the audit
constraint matrix corresponding to the lower bound for m1.}

\item{ub0seq}{integer vector, indicates which rows in the audit
constraint matrix corresponding to the upper bound for m0.}

\item{ub1seq}{integer vector, indicates which rows in the audit
constraint matrix corresponding to the upper bound for m1.}

\item{mono0seq}{integer vector, indicates which rows in the audit
constraint matrix corresponding to the monotonicity conditions
for m0.}

\item{mono1seq}{integer vector, indicates which rows in the audit
constraint matrix corresponding to the monotonicity conditions
for m1.}

\item{monoteseq}{integer vector, indicates which rows in the audit
constraint matrix corresponding to the monotonicity conditions
for the mte.}

\item{mbmap}{integer vector, indexes the X-value associated with
each row in the audit constraint matrix.}
}
\value{
The audit grid is represented using a set of constraint
    matrices. Each point in the audit grid corresponds to a set of
    rows in the constraint matrices. The function simply returns
    the vector of row numbers for the points from the audit grid
    whose corresponding constriants should be added to the original
    LP problem (i.e. the points to add to the original grid).
}
\description{
This function selects which points from the audit grid should be
included into the original grid. Both the initial grid and audit
grid are represented as constraints in an LP problem. This function
selects which points in the audit grid (i.e. which rows in the
audit constraint matrix) should be added to the initial grid
(i.e. should be appended to the initial constraint matrix).
}
