% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_splines.R
\name{sOls}
\alias{sOls}
\title{IV-like weighting function, OLS specifications}
\usage{
sOls(x = NULL, d, j, exx)
}
\arguments{
\item{x}{vector, the value of the covariates other than the
intercept and the treatment indicator.}

\item{d}{0 or 1, indicating treatment or control.}

\item{j}{scalar, position of the component one is interested in
constructing the IV-like weight for.}

\item{exx}{matrix corresponding to E[XX'].}
}
\value{
scalar.
}
\description{
IV-like weighting function for OLS specifications.
}
