% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{sOls1d}
\alias{sOls1d}
\title{IV-like weighting function, OLS specification 1}
\usage{
sOls1d(d, exx)
}
\arguments{
\item{d}{0 or 1, indicating treatment or control.}

\item{exx}{the matrix E[XX']}
}
\value{
scalar.
}
\description{
IV-like weighting function for OLS specification 1.
}
