% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_splines.R
\name{splineInt}
\alias{splineInt}
\title{Integrating splines}
\usage{
splineInt(ub, lb, knots, degree, intercept = FALSE)
}
\arguments{
\item{ub}{scalar, upperbound of integral.}

\item{lb}{scalar, lowerbound of integral.}

\item{knots}{vector, knots of the spline.}

\item{degree}{scalar, degre of spline.}

\item{intercept}{boolean, set to TRUE if spline basis should
include a component so that the basis sums to 1.}
}
\value{
vector, each component being the integral of a basis.
}
\description{
This function simply integrates the splines.
}
