% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{splitIntegrate}
\alias{splitIntegrate}
\title{Function to perform numerical integration over partitions intervals}
\usage{
splitIntegrate(FUN, splits, subdivisions, rel.tol, abs.tol)
}
\arguments{
\item{FUN}{function to be integrated.}

\item{splits}{vector, list of points of discontinuity. This vector
must also include the end points for the whole region of
integration.}
}
\value{
scalar.
}
\description{
Numerical integration of discontinuous functions is very slow. As a
faster alternative, this function takes in a list of points of
discontinuity, partitions the region of integration accordingly,
and performs the integrals over each continuous region. The
integrals are then summed up.
}
