% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{statusString}
\alias{statusString}
\title{Convert status code to string}
\usage{
statusString(status, solver)
}
\arguments{
\item{status}{Status code.}

\item{solver}{Name of solver, either 'gurobi', 'cplexapi', or
'lpsolveapi'.}
}
\value{
Status specific to solver, e.g. 'OPTIMAL (2)'.
}
\description{
This function returns the status code specific to a solver.
}
