% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{sTsls}
\alias{sTsls}
\title{IV-like weighting function, TSLS specification}
\usage{
sTsls(z, j, exz, pi)
}
\arguments{
\item{z}{vector, the value of the instrument.}

\item{j}{scalar, position of the component one is interested in
constructing the IV-like weight for.}

\item{exz}{the matrix E[XZ']}

\item{pi}{the matrix E[XZ']E[ZZ']^{-1}}
}
\value{
scalar.
}
\description{
IV-like weighting function for TSLS specification.
}
