% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_callcheck.R
\name{subsetclean}
\alias{subsetclean}
\title{Auxiliary function: remove extraneous spaces}
\usage{
subsetclean(string)
}
\arguments{
\item{string}{the string object to be cleaned.}
}
\value{
a string
}
\description{
Auxiliary function to remove extraneous spaces from strings.
}
