% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{summary.ivmte}
\alias{summary.ivmte}
\title{Summarize results}
\usage{
\method{summary}{ivmte}(object, ...)
}
\arguments{
\item{object}{an object returned from '\code{ivmte}'.}

\item{...}{additional arguments.}
}
\value{
summarized results.
}
\description{
This function uses the summary method on the ivmte return list.
}
