% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{symat}
\alias{symat}
\title{Generate symmetric matrix}
\usage{
symat(values)
}
\arguments{
\item{values}{vector, the values that enter into the symmetric
matrix. Dimensions will be determined automatically.}
}
\value{
matrix.
}
\description{
Function takes in a vector of values, and constructs a symmetric
matrix from it. Diagonals must be included. The length of the
vector must also be consistent with the number of "unique" entries
in the symmetric matrix. Note that entries are filled in along the
columns (i.e. equivalent to byrow = FALSE).
}
