% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monobound.R
\name{tukeydist}
\alias{tukeydist}
\title{Obtain Tukey half-space quantiles}
\usage{
tukeydist(x, data)
}
\arguments{
\item{x}{vector of values.}

\item{data}{a matrix of data, characterizing the empirical
distribution.}
}
\value{
scalar, representing the quantile of \code{x} under the
    empirical distribution characterized by \code{data}.
}
\description{
This function calculates the Tukey half-space quantiles for
multivariate random variables.
}
