% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wweights.R
\name{unstring}
\alias{unstring}
\title{Auxiliary function that converts a vector of strings into an
expression containing variable names.}
\usage{
unstring(vector)
}
\arguments{
\item{vector}{Vector of variable names (strings).}
}
\value{
An expression for the list of variable names that are not
    strings.
}
\description{
Auxiliary function that converts a vector of strings into an
expression containing variable names.
}
\examples{
\dontrun{
ivmte::unstring(c("a", "b"))
}
}
