% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{vecextract}
\alias{vecextract}
\title{Auxiliary function: extracting elements from strings}
\usage{
vecextract(vector, position, truncation = 0)
}
\arguments{
\item{vector}{the vector from which we want to extract the
elements.}

\item{position}{the position in \code{vector} to extract.}

\item{truncation}{the number of characters from the front of the
element being extracted that should be dropped.}
}
\value{
A chracter/string.
}
\description{
This auxiliary function extracts the (string) element in the
\code{position} argument of the \code{vector} argument.
}
