% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignetteResult}
\alias{vignetteResult}
\title{Vignette results}
\format{An object of class \code{list} of length 31.}
\usage{
vignetteResult
}
\description{
The results for the vignette example. This is to avoid having to
re-run the code when compiling the vignette.
}
\keyword{datasets}
