% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweights.R
\name{wald}
\alias{wald}
\title{Wald weights}
\usage{
wald(D, Z)
}
\arguments{
\item{D}{Matrix, one column being 1s, the other being treatment
indicator D.}

\item{Z}{Matrix, one column being 1s, the other being the
instrument dummy.}
}
\value{
A list of two vectors: one is the weight for D = 0, the
    other is the weight for D = 1.
}
\description{
Function generating the S-weighs for Wald estimand. The function
operates by using only a single dummy as the instrument. It is up
to the user to construct this dummy and subset the data accordingly
to get the correct Wald weights.
}
