% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wweights.R
\name{wate1}
\alias{wate1}
\title{Target weight for ATE}
\usage{
wate1(data)
}
\arguments{
\item{data}{\code{data.frame} on which the estimation is performed.}
}
\value{
The bounds of integration over unobservable \code{u}, as
    well as the multiplier in the weight.
}
\description{
Function generates the target weight for the ATE.
}
