% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_splines.R
\name{wAtt}
\alias{wAtt}
\title{Target weighting function, for ATT}
\usage{
wAtt(z, d, ed)
}
\arguments{
\item{z}{vector, the value of the instrument (redundant in this
function; included to exploit apply()).}

\item{d}{0 or 1, indicating treatment or control (redundant in this
function; included to exploit apply()).}

\item{ed}{scalar, unconditional probability of taking up treatment.}
}
\value{
scalar.
}
\description{
Target weighting function, for the ATT.
}
