% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wweights.R
\name{watt1.mst}
\alias{watt1.mst}
\title{Target weight for ATT}
\usage{
watt1.mst(data, expd1, propensity)
}
\arguments{
\item{data}{\code{data.frame} on which the estimation is performed.}

\item{expd1}{Scalar, the probability that treatment is recieved.}

\item{propensity}{Vector of propensity to take up treatment.}
}
\value{
The bounds of integration over unobservable \code{u}, as
    well as the multiplier in the weight.
}
\description{
Function generates the target weight for the ATT.
}
