% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wweights.R
\name{watu1.mst}
\alias{watu1.mst}
\title{Target weight for ATU}
\usage{
watu1.mst(data, expd0, propensity)
}
\arguments{
\item{data}{\code{data.frame} on which the estimation is performed.}

\item{propensity}{Vector of propensity to take up treatment.}

\item{expd1}{Scalar, the probability that treatment is not
recieved.}
}
\value{
The bounds of integration over unobservable \code{u}, as
    well as the multiplier in the weight.
}
\description{
Function generates the target weight for the ATT.
}
