% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{weights}
\alias{weights}
\title{Generating splines weights}
\usage{
weights(x, knots, i, order)
}
\arguments{
\item{x}{vector, the values at which to evaluate the basis
function.}

\item{knots}{vector, the internal knots.}

\item{i}{integer, the basis component to be evaluated.}

\item{order}{integer, the order of the basis. Do not confuse this
with the degree of the splines, i.e. order = degree + 1.}
}
\value{
scalar.
}
\description{
This function generates the weights required to construct splines
of higher order. This function was coded in accordance to Carl de
Boor's set of notes on splines, "B(asic)-Spline Basics".
}
