% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wweights.R
\name{wgenlate1.mst}
\alias{wgenlate1.mst}
\title{Target weight for generalized LATE}
\usage{
wgenlate1.mst(data, ulb, uub)
}
\arguments{
\item{data}{\code{data.frame} on which the estimation is performed.}

\item{ulb}{Numeric, lower bound of interval.}

\item{uub}{Numeric, upper bound of interval.}
}
\value{
The bounds of integration over unobservable \code{u}, as
    well as the multiplier in the weight.
}
\description{
Function generates the target weight for the generalized LATE,
where the user can specify the interval of propensity scores
defining the compliers.
}
