% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{whichforlist}
\alias{whichforlist}
\title{Auxiliary function: \code{which} for lists}
\usage{
whichforlist(vector, obj)
}
\arguments{
\item{vector}{the vector for which we want to check the entries of}

\item{obj}{the value for which we want the vector to match on.}
}
\value{
a vector of positions where the elements in \code{vector}
    are equal to \code{obj}.
}
\description{
Auxiliary function that makes it possible to use \code{which} with
a list.
}
