% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AE}
\alias{AE}
\title{AngristEvans Data}
\format{
A data frame with 209,133 rows and 5 columns.
\describe{
  \item{worked}{indicator for whether worked in the previous year}
  \item{hours}{weekly hours worked in the previous year}
  \item{morekids}{indicator for having more than two children vs. exactly two children.}
  \item{samesex}{indicator for the first two children having the same sex (male-male or female-female)}
  \item{yob}{the year the woman was born}
}
}
\source{
Derived from Angrist and Evans (1998, The American Economic Review).
}
\usage{
AE
}
\description{
AngristEvans Data
}
\keyword{datasets}
