% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{audit.mst}
\alias{audit.mst}
\title{Audit procedure}
\usage{
audit.mst(data, uname, m0, m1, splinesobj, vars_mtr, terms_mtr, grid.Nu = 20,
  grid.Nx = 50, audit.Nx = 5, audit.Nu = 3, audit.max = 5,
  audit.tol = 1e-08, m1.ub, m0.ub, m1.lb, m0.lb, mte.ub, mte.lb, m0.dec,
  m0.inc, m1.dec, m1.inc, mte.dec, mte.inc, sset, gstar0, gstar1, obseq.tol)
}
\arguments{
\item{data}{\code{data.frame} used to estimate the treatment
effects.}

\item{uname}{name declared by user to represent the unobservable
term in the MTRs.}

\item{m0}{one-sided formula for marginal treatment response
function for control group. The unobservable term can be
entered in using splines.}

\item{m1}{one-sided formula for marginal treatment response
function for treated group.}

\item{grid.Nu}{number of evenly spread points in the interval [0,
1] of the unobservable u used to form the grid for imposing
shape restrictions on the MTRs.}

\item{grid.Nx}{number of evenly spread points of the covariates to
use to form the grid for imposing shape restrictions on the
MTRs.}

\item{audit.Nx}{number of points on the covariates space to audit
in each iteration of the audit procedure.}

\item{audit.max}{maximum number of iterations in the audit
procedure.}

\item{audit.tol}{tolerance for determining when to end the audit
procedure. Namely, if the percentage change in the upper and
lower bounds both fall below \code{audit.tol} between
iterations of the audit, the audit precodure ends.}

\item{m1.ub}{numeric value for upper bound on MTR for treated
group.}

\item{m0.ub}{numeric value for upper bound on MTR for control
group.}

\item{m1.lb}{numeric value for lower bound on MTR for treated
group.}

\item{m0.lb}{numeric value for lower bound on MTR for control
group.}

\item{mte.ub}{numeric value for upper bound on treatment effect
paramter of interest.}

\item{mte.lb}{numeric value for lower bound on treatment effect
paramter of interest.}

\item{m0.dec}{logical, equal to TRUE if we want MTR for control
group to be weakly monotone decreasing.}

\item{m0.inc}{logical, equal to TRUE if we want MTR for control
group to be weakly monotone increasing.}

\item{m1.dec}{logical, equal to TRUE if we want MTR for treated
group to be weakly monotone decreasing.}

\item{m1.inc}{logical, equal to TRUE if we want MTR for treated
group to be weakly monotone increasing.}

\item{mte.dec}{logical, equal to TRUE if we want the MTE to be
weakly monotone decreasing.}

\item{mte.inc}{logical, equal to TRUE if we want the MTE to be
weakly monotone decreasing.}

\item{sset}{a list containing the point estimates and gamma
components associated with each element in the S-set.}

\item{gstar0}{set of expectations for each terms of the MTR for the
control group.}

\item{gstar1}{set of expectations for each terms of the MTR for the
control group.}

\item{obseq.tol}{tolerance level for how much more the solution is
permitted to violate observational equivalence of the IV-like
estimands. The threshold multiplies the violation of the
observational equivalence, i.e. a threshold of 0 corresponds to
the assumption that the model is correctly specified, and that
there is no violation of observational equivalence.}

\item{audit.Nx}{number of points in the interval [0, 1],
corresponding to the normalized value of the unobservable term,
to audit in each iteration of the audit procedure.}
}
\value{
a list. Included in the list is the minimum violation of
    observational equivalence of the set of IV-like estimands, as
    well as the list of matrices and vectors associated with
    solving the LP problem.
}
\description{
This is the wrapper for running the entire audit procedure. This
function sets up the LP problem of minimizing the violation of
observational equivalance for the set of IV-like estimands, while
satisfying boundedness and monotonicity constraints declared by the
user. Rather than enforce boundedness and monotonicity hold across
the entire support of covariates and unobservables, this procedure
enforces the conditions over a subset of points in a grid. This
grid coresponds to the set of values the covariates can take, and a
subset of values of the unobservable term. The size of this grid is
specified by the user in the function arguments. The procedure then
goes on to check whether the constraints are satisfied at points
off the grid. Any point where either the boundedness or
monotonicity constriants are violated are incoporated into the
grid, and the process is repeated until the grid incorporates the
entire support of the covariates, or until some a maximum number of
iterations is reached.
}
