% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\title{Generating design matrices}
\usage{
design(formula, data, subset)
}
\arguments{
\item{formula}{Formula with which to generate the design matrix.}

\item{data}{\code{data.frame} with which to generate the design
matrix.}

\item{subset}{Condition to select subset of data.}
}
\value{
Three matrices are returned: one for the outcome variable,
    Y; one for the second stage covariates, X; and one for the
    first stage covariates, Z.
}
\description{
This function generates the design matrix given an IV
specification.
}
\examples{
design(formula = ey ~ d | z,
           data = dtm,
           subset = z \%in\% c(1, 2))

}
