% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtcf}
\alias{dtcf}
\title{Covariates data set}
\format{A data frame with 10,000 rows and 9 columns.
\describe{
  \item{x1}{covariate 1}
  \item{x2}{covariate 2}
  \item{z1}{instrument 1}
  \item{z2}{instrument 2}
  \item{p}{probability of purchasing mosquito net}
  \item{d}{indicator for treatment (d = 1) versus control (d = 0) group}
  \item{ey0}{counterfactual outcome when not a recipient of treatment}
  \item{ey1}{counterfactual outcome when a recipient of treatment}
  \item{ey}{the observed outcome}
}}
\source{
Simulated.
}
\usage{
dtcf
}
\description{
A simulated population-level data set characterizing the effect of
a treatment on an outcome. The data includes a treatment indicator,
two covariates and two instruments.
}
\keyword{datasets}
