% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{gensset.mst}
\alias{gensset.mst}
\title{Generating LP moments for IV-like estimands}
\usage{
gensset.mst(data, sset, sest, splinesobj, pmodobj, pm0, pm1, ncomponents,
  scount, subset_index, means = TRUE, yvar, dvar, noisy = TRUE)
}
\arguments{
\item{data}{\code{data.frame} used to estimate the treatment
effects.}

\item{sset}{A list, which is modified and returned as the output.}

\item{sest}{A list containing the point estimates and S-weights
corresponding to a particular IV-like estimand.}

\item{splinesobj}{list of spline components in the MTRs for treated
and control groups. Spline terms are extracted using
\code{\link{removeSplines}}.}

\item{pmodobj}{A vector of propensity scores.}

\item{pm0}{A list of the monomials in the MTR for d = 0.}

\item{pm1}{A list of the monomials in the MTR for d = 1.}

\item{ncomponents}{The number of components from the IV regression
we want to include in the S-set.}

\item{scount}{A counter for the number of elements in the S-set.}

\item{subset_index}{An index for the subset of the data the IV
regression is restricted to.}

\item{means}{boolean, set to \code{TRUE} by default. If set to
\code{TRUE}, then the gamma moments are returned, i.e. sample
averages are taken. If set to \code{FALSE}, then no sample
averages are taken, and a matrix is returned. The sample
average of each column of the matrix corresponds to a
particular gamma moment.}

\item{yvar}{name of outcome variable. This is only used if
\code{means = FALSE}, which occurs when the user believes the
treatment effect is point identified.}

\item{dvar}{name of treatment indicator. This is only used if
\code{means = FALSE}, which occurs when the user believes the
treatment effect is point identified.}

\item{noisy}{boolean, default set to \code{TRUE}. If \code{TRUE},
then messages are provided throughout the estimation
procedure. Set to \code{FALSE} to suppress all messages,
e.g. when performing the bootstrap.}
}
\value{
A list containing the point estimate for the IV regression,
    and the expectation of each monomial term in the MTR.
}
\description{
This function takes in the IV estimate and its IV-like
specification, and generates a list containing the corresponding
point estimate, and the corresponding moments (gammas) that will
enter into the constraint matrix of the LP problem.
}
