% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{gmmEstimate}
\alias{gmmEstimate}
\title{GMM estimate of TE under point identification}
\usage{
gmmEstimate(sset, gstar0, gstar1, itermax = 2, tol = 1e-08, obsComp0,
  obsComp1, commonM1Terms, commonM1STerms, uexp0, uexp1, ate = FALSE,
  negate = TRUE, noisy = TRUE)
}
\arguments{
\item{sset}{a list of lists constructed from the function
\link{gensset.mst}. Each inner list should include a
coefficient corresponding to a term in an IV specification, a
matrix of the estimates of the gamma moments conditional on (X,
Z) for d = 0, and a matrix of the estimates of the gamma
moments conditional on (X, Z) for d = 1. The column means of
the last two matrices is what is used to generate the gamma
moments.}

\item{gstar0}{vector, the target gamma moments for d = 0.}

\item{gstar1}{vector, the target gamma moments for d = 1.}

\item{itermax}{integer, maximum number of iterations allowed in the
iterative GMM process. By default this is set to 2 (two-step
GMM).}

\item{tol}{tolerance level for iterative GMM to terminate.}

\item{obsComp0}{vector of observable variables. This is the
observable component of each term in m0.}

\item{obsComp1}{vector of observable variables. This is the
observable component of each term in m1.}

\item{commonM1Terms}{character, vector of variable names of m1 that
are also included in m0. Depending on the target parameter
weight, the error in estimating the target wight of these
variables may need to be excluded from the GMM procedure.}

\item{commonM1STerms}{list, each element is a character vector
corresponding to a spline in m1 also found in m0. The character
vector stores the variable names that interact with a given
spline in both m0 and m1. Depending on the target parameter
weight, the error in estimating the target wight of these
variables may need to be excluded from the GMM procedure.}

\item{uexp0}{vector of exponential on unobservable term in m0. The
values only apply to non-spline terms, though.}

\item{uexp1}{vector of exponential on unobservable term in m1. The
values only apply to non-spline terms, though.}

\item{noisy}{boolean, default set to \code{TRUE}. If \code{TRUE},
then messages are provided throughout the estimation
procedure. Set to \code{FALSE} to suppress all messages,
e.g. when performing the bootstrap.}
}
\value{
a list containing the point estimate of the treatment
    effects, the standard errors, the 90% and 95% confidence
    intervals, the convergence code (see
    \code{\link[stats]{optim}}), the coefficients on the MTR, and
    the variance/covariance matrix of the MTR coefficient
    estimates.
}
\description{
If the user sets the argument \code{point = TRUE} in the function
\code{ivmte}, then it is assumed that the treatment effect
parameter is point identified. The observational equivalence
condition is then set up as a GMM problem. Solving this GMM problem
recovers the coefficients on the MTR functions m0 and m1. Combining
these coefficients with the target gamma moments allows us to
estimate the target treatment effect.
}
