% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lpSetupInfeasible}
\alias{lpSetupInfeasible}
\title{Configure LP environment for diagnostics}
\usage{
lpSetupInfeasible(env, sset)
}
\arguments{
\item{env}{The LP environment}

\item{sset}{List of IV-like estimates and the corresponding gamma
terms.}
}
\value{
Nothing, as this modifies an environment variable to save
    memory.
}
\description{
This function separates the shape constraints from the LP
environment. That way, the model can be solved without any shape
constraints, which is the primary cause of infeasibility. This is
done in order to check which shape constraints are causing the
model to be infeasible. The LP model must be passed as an
environment variable, under the entry \code{$lpobj}. See
\code{\link{lpSetup}}.
}
