% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{removeSplines}
\alias{removeSplines}
\title{Separating splines from MTR formulas}
\usage{
removeSplines(formula)
}
\arguments{
\item{formula}{the formula that is to be parsed.}
}
\value{
a list containing two objects. One object is \code{formula}
    but with the spline components removed. The second object is a
    list. The name of each element is the "uSplines()" command, and
    the elements are a vector of the names of covariates that were
    interacted with the "uSplines()" command.
}
\description{
This function separates out the function call "uSplines()"
potentially embedded in the MTR formulas from the rest of the
fomrula. The terms involving splines are treated separately from
the terms that do not involve splines when creating the gamma
moments.
}
