% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{rescaleX}
\alias{rescaleX}
\title{Function to implement rescaling procedure}
\usage{
rescaleX(sset, dVec, drY, drN)
}
\arguments{
\item{sset}{a list of lists constructed from the function
\link{genSSet}. In the case of a direct regression, 'sset'
contains only one inner list. This list contains the gamma
moment at the individual level.}

\item{dVec}{Vector of treatment statuses from the data.}

\item{drY}{Vector of outcomes from the data.}

\item{drN}{Scalar, number of observations in the data.}
}
\value{
List of rescaled covariates.
}
\description{
This function rescales the matrix of covariates used in the direct
regression to improve the conditioning number and the stability of
the estimation procedure.
}
