% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_covariates.R
\name{s.wald}
\alias{s.wald}
\title{IV-like weighting function, Wald specification}
\usage{
s.wald(z, p.to, p.from, e.to, e.from)
}
\arguments{
\item{z}{vector, the value of the instrument.}

\item{p.to}{P[Z = z'], where z' is value of the instrument the
agent is switching to.}

\item{p.from}{P[Z = z], where z is the value of the instrument the
agent is switching from.}

\item{e.to}{E[D | Z = z'], where z' is the value of the instrument
the agent is switching to.}

\item{e.from}{E[D | Z = z], where z is the value of the instrument the
agent is switching from.}
}
\value{
scalar.
}
\description{
IV-like weighting function for OLS specification 2.
}
