% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunctions_splines.R
\name{sTsls}
\alias{sTsls}
\title{IV-like weighting function, TSLS specification}
\usage{
sTsls(z, d, j, exz, pi)
}
\arguments{
\item{z}{vector, the value of the instrument.}

\item{d}{0 or 1, indicating treatment or control (redundant in this
function; included to exploit apply()).}

\item{j}{scalar, position of the component one is interested in
constructing the IV-like weight for.}

\item{exz}{matrix, corresponds to E[XZ'].}

\item{pi}{matrix, corresponds to E[XZ']E[ZZ']^{-1}, the first stage
regression.}
}
\value{
scalar.
}
\description{
IV-like weighting function for TSLS specification.
}
